Add initial OpenLab charter
This is an initial version of a charter for OpenLab. It is not expected
to be exhausted or a final version by any stretch. The hope is this will
give us a starting point to further refine to make sure we are clearly
declaring how OpenLab works and what is expected from the project.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>